+++ Sources/Spritz/Extensions/String+Extensions.swift
@@ -14,6 +14,20 @@ internal extension String {
   
   /// Returns a copy of the string which is stripped from all whitespaces, special characters and accents.
   func strippedForCF() -> String {
-    self.filter { $0.isLetter }.folding(options: .diacriticInsensitive, locale: .current)
+    self.filter {
+      $0.isLetter
+    }
+    .folding(options: .diacriticInsensitive, locale: nil)
+  }
+}
+
+internal extension Spritz.CodiceFiscale {
+  func strippedChecksum() -> Spritz.CodiceFiscale {
+    assert(count == 16)
+    
+    var mutable = self
+    mutable.removeLast()
+    
+    return mutable
   }
 }