+++ Sources/Spritz/Models/Sex.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+extension Spritz.Models {
+  /// The legal sex of the person.
+  public enum Sex: String {
+    /// A legally male person.
+    case male
+    
+    /// A legally female person.
+    case female
+  }
+}